# specs/001-overtime-management/contracts/api-spec.yaml
openapi: 3.1.0
info:
  title: Minhas Horas API
  version: 1.0.0
  description: REST API for Overtime Management, Time Bank Calculation, Offline Synchronization, and Multi-Format Reporting.
servers:
  - url: /api/v1
    description: Base API prefix
paths:
...
  /compensations:
    get:
      summary: List scheduled and completed compensations
      operationId: listCompensations
      responses:
        '200':
          description: List of compensations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CompensationSchedule'
    post:
      summary: Pre-schedule a compensation
      operationId: createCompensation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCompensationRequest'
      responses:
        '201':
          description: Created compensation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompensationSchedule'
  /compensations/{id}:
    put:
      summary: Update compensation status or hours
      operationId: updateCompensation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCompensationRequest'
      responses:
        '200':
          description: Updated compensation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompensationSchedule'
...
  /reports/export:
    get:
      summary: Export time records and time bank report
      operationId: exportReport
      parameters:
        - name: format
          in: query
          required: true
          schema:
            type: string
            enum: [csv, xlsx, pdf]
        - name: start_date
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: File download stream
          content:
            text/csv:
              schema:
                type: string
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
            application/pdf:
              schema:
                type: string
                format: binary
        '400':
          description: Invalid date range or empty period
components:
  schemas:
...
    CompensationSchedule:
      type: object
      required:
        - id
        - planned_date
        - scheduled_minutes
        - status
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
        planned_date:
          type: string
          format: date
          example: '2026-10-15'
        scheduled_minutes:
          type: integer
          example: 240
        actual_minutes:
          type: integer
          nullable: true
        status:
          type: string
          enum: [Scheduled, Completed, Cancelled]
        notes:
          type: string
        client_updated_at:
          type: string
          format: date-time
    CreateCompensationRequest:
      type: object
      required:
        - id
        - planned_date
        - scheduled_minutes
      properties:
        id:
          type: string
          format: uuid
        planned_date:
          type: string
          format: date
        scheduled_minutes:
          type: integer
        notes:
          type: string
        client_updated_at:
          type: string
          format: date-time
    UpdateCompensationRequest:
      type: object
      properties:
        planned_date:
          type: string
          format: date
        scheduled_minutes:
          type: integer
        actual_minutes:
          type: integer
        status:
          type: string
          enum: [Scheduled, Completed, Cancelled]
        notes:
          type: string
        client_updated_at:
          type: string
          format: date-time
...